Movie Page update
Large update to movie page. Fetches all relevant data from the database about the movie, the user (whether they are logged in or not), (whether their toke is still valid or not), reviews, and possible reviews from the user.

Also learned to deal with the accumulation of uncancelled asynchronous requests, leading to unexpected behavior and the component getting stuck on loading
By implementing axios.CancelToken and passing it to each asynchronous function, I ensured that requests could be canceled when the component unmounted or updated, preventing the accumulation of unresolved requests and resolving the loading issue